✨feat: 관리 영양소 필드명 및 타입 수정

+++ android-app/app/src/main/java/com/swallaby/foodon/domain/main/model/NutrientManage.kt
@@ -8,8 +8,8 @@ data class NutrientManage(
     val nutrientCode: NutrientCode = NutrientCode.CARBS,
     val unit: UnitType = UnitType.GRAM,
     val intake: Double = 0.0,
-    val minRecommend: Int = 0,
-    val maxRecommend: Int = 0,
+    val minRecommend: Double = 0.0,
+    val maxRecommend: Double = 0.0,
     val healthEffect: HealthEffect = HealthEffect.UNKNOWN,
     val status: ManageStatus = ManageStatus.ADEQUATE
 )
